Revert "조건 캡슐화 + 이름 붙이기, constant 분리"
This reverts commit ce54d3201fe5ba3a609ee23c88ff6ff8ccc87d83.
